+++ Changes.md
@@ -1,3 +1,4 @@
+* Fix saving of subjective data to local storage 
 * Keep paragraph breaks entered into awards script entry
 * Allow tournament levels to be renamed
 * fix error editing team data